backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
aiofiles>=23.2.1
pydantic>=2.0.0
pydantic-settings>=2.2.0
authlib>=1.3.0
httpx>=0.27.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
